backend/requirements.txt
aenum==3.1.11
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
dill==0.3.5.1
Django==3.2.15
executing==1.0.0
fibheap==0.2.1
fonttools==4.37.2
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonpickle==2.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mpmath==1.2.1
networkx==2.6.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parso==0.8.3
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
ply==3.11
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pyvis==0.1.9
qiskit-ignis==0.7.0
qiskit-terra==0.21.2
# qntsim==0.1.1
qutip==4.6.3
retworkx==0.11.0
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
stevedore==3.5.0
symengine==0.9.2
sympy==1.10.1
tabulate==0.8.9
traitlets==5.4.0
tweedledum==1.1.1
wcwidth==0.2.5
psycopg2
qiskit-aer
djangorestframework==3.13.1
django-cors-headers
djangorestframework-simplejwt==5.1.0
joblib
gunicorn
websocket-client